stylish-haskell: fix code stylish-haskell fails on
The tx-generator code involved a definition whose definition needed to
be moved fully within an ifdef. The second involved what appeared to be
a string spanning lines without the newline escaped in a branch of an
ifdef seemingly not usually seen. This version of this fix is informed by
having read the Formatting hackage documentation and is corrected
relative to prior ones, though dealing with the ifdefs is awkward enough
that the branch of the ifdef in question has still not been directly
compile-tested by me.