locli-latex: anomaly comparison table
I got things working to a degree that I'm afraid of losing my progress,
so I'm saving a snapshot of an intermediate state. The basic thing is
that I fixed up the generation of one of the tables to handle
comparisons overall. From there, there is some summary information I
need to fill in too and to replicate it all across the other tables.