locli-latex: clean up orphan, import warnings
The orphan warning was on deriving Foldable for Interval within CDF.hs
itself. This moves it to Util.hs and adds DerivingFoldable there. The
import warning was importing Data.Aeson into Report.hs. Generic and
Foldable didn't need their deriving clauses for CDFList to be
standalone, so this moves them to be attached to the data declaration.

This should be squashed away entirely.