locli-latex: clean up warnings and whitespace
In CDF.hs, there's an orphan instance for Foldable for Interval, then
CDFList's Generic and Foldable instances need no extra instance
prerequisites, so they can be moved into a deriving clause attached to
the data definition. The Data.Aeson import warning was unused, so that
just got removed. The Foldable instance for Interval got moved to
Util.hs and the DerivingFoldable option got set there for it.

This should be squashed.