txgen-mvar: output info in signal handler
This only does additional output upon receipt of a signal. Other
aspects of orderly shutdown need to be handled in other ways. I
see a bug where the spawning thread's ID is what has its weak
reference dereferenced instead of the thread the signal was actually
sent to, so there is more to come.