txgen-mvar: use TVar for Env AsyncBenchmarkControl
In order to thread the AsyncBenchmarkControl through the contexts
surrounding the creation and destruction of the Async structures and
overall container, this stores a TVar (Maybe AsyncBenchmarkControl) as
a value in a Map where previously it was just AsyncBenchmarkControl.
The idea is to use the reference to it to be able to use it in the
context of a signal handler by packaging the reference data with the
code pointer in a partial application or monadic context or similar.
With that data in hand, it's just a matter of iterating over the
threads and reaping them all.