txgen-mvar: label threads, report labels
mkarg@iohk.io wrote the nix code and the beginnings of the Haskell
side. So this commit goes about creating node aliases, represented in
Haskell in the NodeDescription structure, and using them in reporting,
which happens in the exception handlers. The node aliases contribute to
thread labels reported in the exception handlers where possible; older
base versions lack accessors for thread labels.

handleTxSubmissionClientError is the main exception handler involved.
txSubmissionClient, consumeTxsNonBlocking and newTpsThrottle are
spawned as threads within walletBenchmark and labelled accordingly,
with the txSubmissionClient labels further elaborated upon with the
node submitted to. SubmitMode's threadName field for the Benchmark
alternative is produced by the compiler and corresponds to a label for
a set of threads tracked in an AsyncBenchmarkControl structure as
opposed to a label for an individual thread, hence the effort to
generate thread labels by other means.